etl/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
ensure==1.0.4
geopy==2.4.1
numpy==2.0.0
pandas==2.2.2
python_crontab==3.2.0
croniter==2.0.5
PyYAML==6.0.1
Requests==2.32.3
SQLAlchemy==2.0.31
tqdm==4.66.4
psycopg2-binary==2.9.9
jupyter==1.0.0
matplotlib==3.9.1
seaborn==0.13.2